Enable building against macOS 11 SDK
Currently, the newest macOS SDK that Nixpkgs for Darwin will build
against is that for macOS 11.  In order for builds against that SDK to
succeed, APIs that are only available from macOS 12 must be excluded
from the Objective-C source code before the compiler sees it.

This commit does that by using the same #ifdef mechanism currently used
to enable building against the macOS 12 and 12.3 SDKs.